chanstate: group commitment kv codecs
Move the commitment, HTLC, commit-diff, and commitment auxiliary TLV
codecs into a single kv_commitment file. This keeps storage
serialization grouped by the commitment domain while preserving the kv_
prefix for backend-specific on-disk helpers.
