--- chanstate/commitment_serialization.go
@@ -1,39 +1,0 @@
-package chanstate
-
-import "io"
-
-// SerializeChanCommit serializes the channel commitment.
-func SerializeChanCommit(w io.Writer, c *ChannelCommitment) error {
-	if err := WriteElements(w,
-		c.CommitHeight, c.LocalLogIndex, c.LocalHtlcIndex,
-		c.RemoteLogIndex, c.RemoteHtlcIndex, c.LocalBalance,
-		c.RemoteBalance, c.CommitFee, c.FeePerKw, c.CommitTx,
-		c.CommitSig,
-	); err != nil {
-		return err
-	}
-
-	return SerializeHtlcs(w, c.Htlcs...)
-}
-
-// DeserializeChanCommit deserializes the channel commitment.
-func DeserializeChanCommit(r io.Reader) (ChannelCommitment, error) {
-	var c ChannelCommitment
-
-	err := ReadElements(r,
-		&c.CommitHeight, &c.LocalLogIndex, &c.LocalHtlcIndex,
-		&c.RemoteLogIndex, &c.RemoteHtlcIndex, &c.LocalBalance,
-		&c.RemoteBalance, &c.CommitFee, &c.FeePerKw, &c.CommitTx,
-		&c.CommitSig,
-	)
-	if err != nil {
-		return c, err
-	}
-
-	c.Htlcs, err = DeserializeHtlcs(r)
-	if err != nil {
-		return c, err
-	}
-
-	return c, nil
-}
